Stars added, Sections Ready
CSS Classes added:
hr.star-hello,
hr.star-darkbkg,
hr.star-bluebkg,
hr.star-hello-light

You not only have to modify the CSS in freelancer.css, but also minify
it and add it to freelancer.min.css in order for the stars to be
formatted properly.

The “border-color” is the lines. The “background color” gives the
illusory appearance of a gap behind the stars, and should be the same
color as the section that the star is in. “Color” is simply the color
of the star, which should be the same as “border-color”.